fix UTs

+++ eheap/eheap_test.go
@@ -97,7 +97,7 @@ func TestSetMin(t *testing.T) {
 	// All timestamps less than 5
 	seen := make(map[int64]bool)
 	for _, item := range removed {
-		require.Less(item.Expiry(), 5)
+		require.Less(item.Expiry(), int64(5))
 		_, ok := seen[item.Expiry()]
 		require.False(ok, "Incorrect item removed.")
 		seen[item.Expiry()] = true
